# Security_API_gateway/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MarketMind AI — Security API Gateway",
    "version": "1.0.0",
    "description": "Unified gateway: authentication, RBAC, rate-limiting, audit logging, and reverse-proxy to the Backend service.\n\nAll protected routes require `Authorization: Bearer <accessToken>`.\n\nTokens are issued by `POST /api/auth/login` (access: 1 h) and refreshed via `POST /api/auth/refresh` (refresh: 7 d)."
  },
  "servers": [
    { "url": "http://localhost:7000", "description": "Local Security API Gateway" }
  ],
  "tags": [
    { "name": "Auth",      "description": "Register, login, refresh, logout, current-user" },
    { "name": "Inventory", "description": "Stock management (proxied to backend)" },
    { "name": "Sales",     "description": "Sales transactions and CSV upload (proxied to backend)" },
    { "name": "Dashboard", "description": "KPI summary (proxied to backend)" },
    { "name": "Analytics", "description": "Revenue and top-product analytics (proxied to backend)" },
    { "name": "Products",  "description": "Product catalogue (proxied to backend)" },
    { "name": "Customers", "description": "Customer list (proxied to backend)" }
  ],
  "paths": {
    "/api/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register a new user",
        "description": "Creates a new user account. Password is bcrypt-hashed (10 rounds). Rate-limited to 10 requests per 15 min per IP.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterRequest" },
              "example": { "name": "Alice Owner", "email": "alice@marketmind.dev", "password": "Password1!", "roleId": 1 }
            }
          }
        },
        "responses": {
          "201": { "description": "User registered successfully", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RegisterResponse" } } } },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "409": { "description": "Email already registered" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login and obtain JWT tokens",
        "description": "Validates credentials and issues a short-lived access token (1 h) and a long-lived refresh token (7 d). Rate-limited to 10 requests per 15 min per IP. Login failures are recorded in the audit log.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" },
              "example": { "email": "alice@marketmind.dev", "password": "Password1!" }
            }
          }
        },
        "responses": {
          "200": { "description": "Login successful", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoginResponse" } } } },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "description": "Invalid credentials" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "tags": ["Auth"],
        "summary": "Refresh access token",
        "description": "Issues a new access token from a valid refresh token. The refresh token itself is not rotated.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RefreshRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "New access token issued", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RefreshResponse" } } } },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "403": { "description": "Invalid or expired refresh token" }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": ["Auth"],
        "summary": "Get current user profile",
        "description": "Returns the authenticated user's id, name, email, roleId, and role name. Requires a valid access token.",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Current user profile", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MeResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "description": "User not found" }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": ["Auth"],
        "summary": "Logout (stateless)",
        "description": "Stateless logout — the server records the event in the audit log. The client is responsible for discarding its tokens.",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Logged out successfully" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/inventory": {
      "get": {
        "tags": ["Inventory"],
        "summary": "List all inventory with low-stock flags",
        "description": "Returns every inventory record annotated with `lowStock: boolean` and a summary count. Access audited.",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Inventory list + summary", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/InventoryListResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
    "/api/inventory/low-stock": {
      "get": {
        "tags": ["Inventory"],
        "summary": "List only low-stock items",
        "description": "Returns only items whose `quantity <= lowStockThreshold`, ordered by quantity ascending. Access audited.",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Low-stock items with count summary" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
    "/api/inventory/add": {
      "post": {
        "tags": ["Inventory"],
        "summary": "Increment stock for a product",
        "description": "Increments the stock quantity for the given product. Requires `quantity >= 1`. Write ops rate-limited to 30 per 15 min. Access audited.",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/InventoryWriteRequest" } } }
        },
        "responses": {
          "200": { "description": "Stock incremented" },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "description": "Product not found" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
    "/api/inventory/update": {
      "put": {
        "tags": ["Inventory"],
        "summary": "Set absolute stock level for a product",
        "description": "Sets the stock quantity to the supplied value. Requires `quantity >= 0`. Write ops rate-limited to 30 per 15 min. Access audited.",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/InventoryUpdateRequest" } } }
        },
        "responses": {
          "200": { "description": "Stock updated" },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "description": "Product not found" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
    "/api/inventory/delete": {
      "delete": {
        "tags": ["Inventory"],
        "summary": "Delete inventory record for a product",
        "description": "Removes the inventory record for the given product. Write ops rate-limited to 30 per 15 min. Access audited.",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "productCode": { "type": "string", "example": "PROD-0001" } },
                "required": ["productCode"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Inventory record deleted" },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "description": "Product not found" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
    "/api/sales": {
      "get": {
        "tags": ["Sales"],
        "summary": "List sales transactions (paginated, filterable)",
        "description": "Query parameters are Joi-validated at the gateway before forwarding. All query params are optional. Access audited.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "page",       "in": "query", "schema": { "type": "integer", "default": 1,  "minimum": 1 },          "description": "Page number (1-based)" },
          { "name": "pageSize",   "in": "query", "schema": { "type": "integer", "default": 20, "minimum": 1, "maximum": 100 }, "description": "Results per page" },
          { "name": "customerId", "in": "query", "schema": { "type": "string", "format": "uuid" },                       "description": "Filter by customer UUID" },
          { "name": "productId",  "in": "query", "schema": { "type": "string", "format": "uuid" },                       "description": "Filter by product UUID" },
          { "name": "startDate",  "in": "query", "schema": { "type": "string", "format": "date", "example": "2024-01-01" }, "description": "Transactions on or after this date" },
          { "name": "endDate",    "in": "query", "schema": { "type": "string", "format": "date", "example": "2024-12-31" }, "description": "Transactions on or before this date (must be >= startDate)" }
        ],
        "responses": {
          "200": { "description": "Paginated list with pagination meta", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SalesListResponse" } } } },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
    "/api/sales/{id}": {
      "get": {
        "tags": ["Sales"],
        "summary": "Get a single sale by ID",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" }, "description": "Sale transaction UUID" }
        ],
        "responses": {
          "200": { "description": "Single sale transaction with customer/product/user details" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "description": "Sale not found" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
    "/api/sales/upload": {
      "post": {
        "tags": ["Sales"],
        "summary": "Upload a CSV file of sales rows",
        "description": "Accepts a `.csv` file (max 10 MB). Required columns: CustomerID, ProductID, Quantity, Price, TransactionDate. Rate-limited to 5 uploads per 15 min. Upload action audited.",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": { "type": "string", "format": "binary", "description": "CSV file with sales rows" }
                },
                "required": ["file"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Upload summary (filename, uploadedAt, recordsInserted, duplicatesRemoved, invalidRows, validationErrors)", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UploadSummaryResponse" } } } },
          "400": { "description": "No file, wrong format, or all rows invalid" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
    "/api/dashboard/summary": {
      "get": {
        "tags": ["Dashboard"],
        "summary": "KPI summary",
        "description": "Returns totalCustomers, totalProducts, totalSales, totalRevenue.",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Dashboard KPI object" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/api/analytics/summary": {
      "get": {
        "tags": ["Analytics"],
        "summary": "Analytics summary — revenue, sales count, top-5 products",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Analytics summary object" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/api/products": {
      "get": {
        "tags": ["Products"],
        "summary": "List all products",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Array of products" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/api/customers": {
      "get": {
        "tags": ["Customers"],
        "summary": "List customers (up to 100, alphabetical)",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Array of customers" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token issued by POST /api/auth/login. Expires in 1 hour."
      }
    },
    "responses": {
      "Unauthorized":    { "description": "Missing or invalid/expired JWT" },
      "Forbidden":       { "description": "Valid JWT but role lacks permission for this resource/method" },
      "ValidationError": { "description": "Joi validation failed — see errors array" },
      "RateLimited":     { "description": "Too many requests — retry after the window resets" }
    },
    "schemas": {
      "RegisterRequest": {
        "type": "object",
        "required": ["name", "email", "password", "roleId"],
        "properties": {
          "name":     { "type": "string",  "maxLength": 100 },
          "email":    { "type": "string",  "format": "email" },
          "password": { "type": "string",  "minLength": 6 },
          "roleId":   { "type": "integer", "enum": [1, 2, 3, 4], "description": "1=Business Owner, 2=Store Manager, 3=Sales Executive, 4=System Administrator" }
        }
      },
      "RegisterResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "message": { "type": "string" },
          "user": {
            "type": "object",
            "properties": {
              "id":     { "type": "string", "format": "uuid" },
              "name":   { "type": "string" },
              "email":  { "type": "string" },
              "roleId": { "type": "integer" }
            }
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email":    { "type": "string", "format": "email" },
          "password": { "type": "string" }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "success":      { "type": "boolean" },
          "message":      { "type": "string" },
          "accessToken":  { "type": "string", "description": "JWT, expires in 1 h" },
          "refreshToken": { "type": "string", "description": "JWT, expires in 7 d" },
          "user": {
            "type": "object",
            "properties": {
              "id":     { "type": "string" },
              "name":   { "type": "string" },
              "email":  { "type": "string" },
              "roleId": { "type": "integer" }
            }
          }
        }
      },
      "RefreshRequest": {
        "type": "object",
        "required": ["refreshToken"],
        "properties": {
          "refreshToken": { "type": "string" }
        }
      },
      "RefreshResponse": {
        "type": "object",
        "properties": {
          "success":     { "type": "boolean" },
          "accessToken": { "type": "string" }
        }
      },
      "MeResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "user": {
            "type": "object",
            "properties": {
              "id":     { "type": "string" },
              "name":   { "type": "string" },
              "email":  { "type": "string" },
              "roleId": { "type": "integer" },
              "role":   { "type": "object", "properties": { "name": { "type": "string" } } }
            }
          }
        }
      },
      "InventoryWriteRequest": {
        "type": "object",
        "required": ["productCode", "quantity"],
        "properties": {
          "productCode": { "type": "string",  "example": "PROD-0001" },
          "quantity":    { "type": "integer", "minimum": 1, "example": 10 }
        }
      },
      "InventoryUpdateRequest": {
        "type": "object",
        "required": ["productCode", "quantity"],
        "properties": {
          "productCode": { "type": "string",  "example": "PROD-0001" },
          "quantity":    { "type": "integer", "minimum": 0, "example": 50 }
        }
      },
      "InventoryListResponse": {
        "type": "object",
        "properties": {
          "success":   { "type": "boolean" },
          "inventory": { "type": "array", "items": { "type": "object" } },
          "summary": {
            "type": "object",
            "properties": {
              "total":         { "type": "integer" },
              "lowStockCount": { "type": "integer" }
            }
          }
        }
      },
      "SalesListResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data":    { "type": "array", "items": { "type": "object" } },
          "pagination": {
            "type": "object",
            "properties": {
              "total":      { "type": "integer" },
              "page":       { "type": "integer" },
              "pageSize":   { "type": "integer" },
              "totalPages": { "type": "integer" }
            }
          }
        }
      },
      "UploadSummaryResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "message": { "type": "string" },
          "summary": {
            "type": "object",
            "properties": {
              "filename":          { "type": "string" },
              "uploadedAt":        { "type": "string", "format": "date-time" },
              "totalRowsRead":     { "type": "integer" },
              "recordsInserted":   { "type": "integer" },
              "duplicatesRemoved": { "type": "integer" },
              "invalidRows":       { "type": "integer" },
              "validationErrors":  {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "row":    { "type": "integer" },
                    "reason": { "type": "string" },
                    "data":   { "type": "object" }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}